Extract ordered window helpers
